Implemented 'map' in terms of 'traverse'
