+++ src/main/scala/chapter_12/Applicative.scala
@@ -101,28 +101,3 @@ object Applicative {
   }
 
 }
-
-trait Monad[F[_]] extends Applicative[F] {
-  def flatMap[A, B](fa: F[A])(f: A => F[B]): F[B]
-
-  def join[A](ffa: F[F[A]]): F[A] = flatMap(ffa)(fa => fa)
-
-  def compose[A, B, C](f: A => F[B], g: B => F[C]): A => F[C] =
-    a => flatMap(f(a))(g)
-
-  override def map[A, B](fa: F[A])(f: A => B): F[B] =
-    flatMap(fa)((a: A) => unit(f(a)))
-
-  def map2[A, B, C](fa: F[A], fb: F[B])(f: (A, B) => C): F[C] =
-    flatMap(fa)(a => map(fb)(b => f(a, b)))
-}
-
-object Monad {
-
-  def eitherMonad[E] = new Monad[({type f[x] = Either[E, x]})#f] {
-    override def unit[A](a: => A): Either[E, A] = Right(a)
-
-    override def flatMap[A, B](fa: Either[E, A])(f: (A) => Either[E, B]): Either[E, B] = fa.flatMap(f)
-  }
-
-}
